requirements.txt
alembic==1.14.0
SQLAlchemy==2.0.36

bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0

fastapi==0.115.5
email_validator==2.2.0
fastapi-cache2==0.2.2

psycopg2-binary==2.9.10

pydantic==2.10.2
pydantic-settings==2.6.1

python-dotenv==1.0.1
python-multipart==0.0.19

redis==5.2.0
slowapi==0.1.9
uvicorn==0.32.1
gunicorn==23.0.0
